requirements.txt
django==2.2.7
gunicorn
djangorestframework